Rename and simplify Title API
